Pull sql pooling library into its own extension; drop postgresql library
