+++ .github/workflows/integrationtests.yaml
@@ -108,32 +108,3 @@ jobs:
         env:
           RUN_INTEGRATION_TEST: true
         run: ./gradlew clean extensions:iam:daps:check
-
-  Lib-Common-SQL-Postgresql-Integration-Test:
-    runs-on: ubuntu-latest
-
-    services:
-      postgres:
-        image: postgres:14.1
-        ports:
-          - 5432:5432
-        env:
-          POSTGRES_PASSWORD: pgpass
-          POSTGRES_USER: pguser
-          POSTGRES_DB: pgdb
-
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up JDK 11
-        uses: actions/setup-java@v1
-        with:
-          java-version: '11'
-      - name: Postgres Connection Factory Test
-        env:
-          RUN_INTEGRATION_TEST: true
-          POSTGRES_PASSWORD: pgpass
-          POSTGRES_USER: pguser
-          POSTGRES_DB: pgdb
-          POSTGRES_HOST: localhost
-          POSTGRES_PORT: 5432
-        run: ./gradlew :common:libraries:sql-postgresql-lib:check
