--- common/libraries/sql-postgresql-lib/README.md
@@ -1,13 +1,0 @@
-# SQL PostgreSQL Library
-
-The PostgreSQL Library implements the `ConnectionFactory` interface for PostgreSQL.
-
-## Usage
-
-### 1. Configuration
-
-Configure the connection factory by creation a new instance of the `PostgreSqlConnectionFactoryConfig`.
-
-### 2. Instantiation
-
-Create a new instance of the `PostgreSqlConnectionFactory` using the `PostgreSqlConnectionFactoryConfig`.
