+++ settings.gradle.kts
@@ -18,10 +18,6 @@ rootProject.name = "dataspaceconnector"
 
 include(":common:util")
 
-include(":common:libraries:sql-lib")
-include(":common:libraries:sql-pool-commons-lib")
-include(":common:libraries:sql-postgresql-lib")
-
 // EDC core modules
 include(":core")
 include(":core:policy:policy-engine")
@@ -119,6 +115,8 @@ include(":extensions:data-plane:data-plane-validation-facade:data-plane-validati
 include(":extensions:data-plane:data-plane-http-proxy:data-plane-http-proxy-provider")
 include(":extensions:data-plane:data-plane-http-proxy:data-plane-http-proxy-consumer")
 include(":extensions:data-plane:data-plane-http-proxy:data-plane-http-proxy-core")
+include(":extensions:sql:common")
+include(":extensions:sql:pool:apache-commons-pool")
 
 // modules for launchers, i.e. runnable compositions of the app
 include(":launchers:basic")
